TST: API: updated tests for datahandler>extraction rename
